client_drf/app/views.py
from django.conf import settings
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated



class UserView(APIView):
    permission_classes = (IsAuthenticated,)

    def get(self, request):
        user = request.user
        print(request.user, request.user.__dict__)
        return Response({'user': request.user.username, 'id': request.user.id})
