requirements.txt
asgiref==3.4.1
Django==4.0.1
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.10.0.0
zope.interface==5.4.0
